+++ R/trakt.movies.trending.R
@@ -4,6 +4,8 @@
 #' According to the API docs, it returns all movies being watched right now, where
 #' movies with the most users are returned first.
 #' @param limit Number of movies to return. Is coerced to \code{integer} and must be greater than 0.
+#' @param page Page to return (default is \code{1})
+#' for \href{http://docs.trakt.apiary.io/#introduction/pagination}{pagination}.
 #' @param extended Whether extended info should be provided.
 #' Defaults to \code{"full"}, can either be \code{"min"} or \code{"full"}
 #' @return A \code{data.frame} containing trending movies with their number of watchers, name and ids
@@ -16,21 +18,22 @@
 #' get_trakt_credentials() # Set required API data/headers
 #' trakt.movies.trending(5)
 #' }
-trakt.movies.trending <- function(limit = 10, extended = "min"){
+trakt.movies.trending <- function(limit = 10, page = 1, extended = "min"){
   limit <- as.integer(limit)
-  if (limit < 1){
-    stop("Limit must be greater than zero")
+  page  <- as.integer(page)
+  if (limit < 1 | page < 1){
+    stop("Limit and page must be greater than zero")
   }
-  ids <- NULL; movie <- NULL
 
   # Construct URL, make API call
   baseURL  <- "https://api-v2launch.trakt.tv/movies/trending"
-  url      <- paste0(baseURL, "?page=1&limit=", limit, "&extended=", extended)
+  url      <- paste0(baseURL, "?page=", page, "&limit=", limit, "&extended=", extended)
   response <- trakt.api.call(url)
 
   # Spreading out ids to get a flat data.frame
-  response$movie <- cbind(subset(response$movie, select = -ids), response$movie$ids)
-  response       <- cbind(subset(response, select = -movie), response$movie)
+  response$movie <- cbind(response$movie[names(response$movie) != "ids"], response$movie$ids)
+  response       <- cbind(response[names(response) != "movie"], response$movie)
+  response       <- convert_datetime(response)
 
   return(response)
 }